pyproject.toml
[project]
name = "short-predictions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "ipykernel>=7.0.1",
    "kaleido>=1.1.0",
    "notebook>=7.4.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "plotly>=6.3.0",
    "pyarrow>=21.0.0",
    "scikit-learn>=1.7.0",
    "torchsummary>=1.5.1",
    "torchview>=0.2.7",
    "torchvision>=0.22.1",
    "traffic>=2.13",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "pip>=25.2",
    "ruff>=0.14.1",
]

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
